% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{comp_list}
\alias{comp_list}
\title{Data frame of competition names and ids}
\usage{
comp_list()
}
\arguments{
\item{none}{}
}
\value{
data frame of competition numbers (in \code{id}) and competition names (in \code{name})
}
\description{
Data frame of competition names and ids
}

