% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{compiled_model}
\alias{compiled_model}
\title{Obtain compiled model}
\usage{
compiled_model()
}
\arguments{
\item{None.}{}
}
\value{
Stan model object for passing into model-fitting and model-describing
  functions.
}
\description{
Get object containing compiled Stan model
}

