% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{connect}
\alias{connect}
\title{connect to database}
\usage{
connect(what = "SQLite", name = "/home/ken/sports/scoresway/soccer.db")
}
\arguments{
\item{what}{type of database (default SQLite)}

\item{name}{name of database (default soccer.db)}
}
\value{
connection
}
\description{
connect to database
}

