% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{do_predict}
\alias{do_predict}
\title{Obtain predictive distribution for teams numbers i and j from fitted model z}
\usage{
do_predict(z, i, j)
}
\arguments{
\item{z}{fitted model from \code{sample_model}}

\item{i}{number (not name) of home team}

\item{j}{number (not name) of away team}
}
\value{
list of: \code{dist}, predictive probabilities of each score,
  \code{prob} posterior predictive probabilities of win, draw and loss for
  team \code{i} at home against team \code{j}.
}
\description{
Obtain predictive distribution for teams numbers i and j from fitted model z
}

