% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{england}
\alias{england}
\title{England soccer data}
\format{A data frame with 349 rows and 4 columns:
\describe{
  \item{t1}{Name of home team}
  \item{t2}{Name of away team}
  \item{s1}{Home team goals}
  \item{s2}{Away team goals}
}}
\usage{
england
}
\description{
A dataset containing all matches between the start of the season and May 4 in the 2014-15 season of the English Premier (soccer) league.
}
\keyword{datasets}

