% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{find_comp}
\alias{find_comp}
\title{Find league by date of last match and name pattern}
\usage{
find_comp(date, pat)
}
\arguments{
\item{date}{Date that matches must be after}

\item{pat}{Pattern to match in league name}
}
\value{
data frame: \code{date} is date of last game in league, \code{comp}
  is competition number, \code{name} is competition name
}
\description{
Find league by date of last match and name pattern
}

