% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_comps}
\alias{get_comps}
\title{get several comps and glue results into one data frame}
\usage{
get_comps(v)
}
\arguments{
\item{v}{vector of competition IDs}
}
\value{
data frame of games like \code{get_comp}
}
\description{
get several comps and glue results into one data frame
}

