% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{last_date}
\alias{last_date}
\title{Date of latest (scheduled) match for each competition}
\usage{
last_date()
}
\arguments{
\item{none}{}
}
\value{
data frame of competition IDs (in \code{comp}) and last date (in \code{date}, as \code{Date})
}
\description{
Date of latest (scheduled) match for each competition
}

