% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{m}
\alias{m}
\title{Small data set for testing}
\format{A data frame with 6 rows and 4 columns:
\describe{
  \item{t1}{Name of home team}
  \item{t2}{Name of away team}
  \item{s1}{Home team goals}
  \item{s2}{Away team goals}
}}
\usage{
m
}
\description{
A dataset consisting of fictitious matches between 3 teams, with team
\code{a} being the strongest and team \code{c} the weakest, with a noticeable
home field advantage
}
\keyword{datasets}

