% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makexy.R
\name{makexy}
\alias{makexy}
\title{Create input to sampler from data frame}
\usage{
makexy(res)
}
\arguments{
\item{res}{Data frame of results with columns team1, team2, score1, score2}
}
\value{
list containing design matrix of teams (as numbers), matrix of scores, names of teams
}
\description{
Create input to sampler from data frame
}
\examples{
makexy(m)
}

