% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{optim.model}
\alias{optim.model}
\title{Posterior mode estimate of Poisson model}
\usage{
optim.model(p.sc, xylist)
}
\arguments{
\item{p.sc}{compiled Poisson object}

\item{xylist}{data processed (fix)}
}
\value{
Stan optimizing object
}
\description{
Obtain posterior mode estimates of ratings
}

