% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poistan.R
\docType{package}
\name{poistan}
\alias{poistan}
\alias{poistan-package}
\title{poistan: a package for fitting Poisson models to goal-scoring data using Stan}
\description{
I  don't know what's going here yet.
}

