% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runmodel.R
\name{sample.model}
\alias{sample.model}
\title{Sample from Poisson model}
\usage{
sample.model(p.sc, xylist, nit = 10000)
}
\arguments{
\item{p.sc}{compiled Poisson object}

\item{xylist}{data processed (fix)}

\item{nit}{number of iterations of sampler}
}
\value{
Stan sampling object
}
\description{
Obtain Stan sampling object
}

