% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\name{show_matches}
\alias{show_matches}
\title{Show all the matches in a competition involving input team}
\usage{
show_matches(comp, name)
}
\arguments{
\item{comp}{matches data frame}

\item{name}{of team to search for}
}
\value{
data frame containing just matches for that team
}
\description{
Show all the matches in a competition involving input team
}

