% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makexy.R
\name{split_half}
\alias{split_half}
\title{Split vector into two halves}
\usage{
split_half(x)
}
\arguments{
\item{x}{vector to split in half}
}
\value{
list of length 2, each component containing half of original vector
}
\description{
Split vector into two halves
}
\examples{
split_half(letters[1:6])
}

