% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_comp}
\alias{get_comp}
\title{get competition from data base}
\usage{
get_comp(id)
}
\arguments{
\item{id}{competition id}
}
\value{
data frame of results
}
\description{
get competition from data base
}

