% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\name{show_rating}
\alias{show_rating}
\title{show mean ratings}
\usage{
show_rating(z)
}
\arguments{
\item{z}{output from sampling (list: `z` of class `stanfit`, `names` vector of team names)}
}
\value{
list containing `r`: table of ratings, `h` home field advantage
}
\description{
offensive, defensive and home field
}

